pyproject.toml
[project]
name = "griot"
version = "0.1.0"
description = "AI-powered media processing API"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi",
    "uvicorn",
    "pydantic",
    "boto3",
    "redis",
    "sqlalchemy",
    "asyncpg",
    "alembic",
    "psycopg2-binary",
    "aiohttp",
    "aiofiles",
    "python-multipart",
    "requests",
    "slowapi",
    "passlib[bcrypt]",
    "bcrypt",
    "python-jose[cryptography]",
    "Pillow",
    "ffmpeg-python",
    "opencv-python>=4.8.0",
    "pytesseract",
    "faster-whisper",
    "yt-dlp",
    "youtube-transcript-api",
    "openai",
    "groq",
    "together",
    "librosa>=0.10.0",
    "webrtcvad-wheels",
    "pydub",
    "srt",
    "transformers>=4.36.0",
    "torch>=2.0.0",
    "scipy>=1.11.0",
    "numpy>=1.24.0",
    "python-dotenv",
    "psutil",
    "loguru",
    "edge-tts==7.2.7",
    "moviepy>=2.0.0",
    "emoji",
    "kokoro-onnx",
    "soundfile",
    "onnxruntime",
    "huggingface_hub",
    "markitdown[all]",
    "python-docx",
    "openpyxl",
    "pdfplumber",
    "beautifulsoup4",
    "lxml",
    "langextract",
    "sse-starlette",
    "pydantic-core",
    "piper-tts",
    "mcp",
    "fastmcp",
    "authlib",
    "python-jose[cryptography]",
    "any-llm-sdk[all]",
    "agno",
    "pgvector",
    "ddgs",
    "duckduckgo-search",
    "yfinance",
    "feedparser",
    "tldextract",
    "backoff",
    "html5lib",
    "langdetect",
    "python-iso639",
    "python-magic",
    "rapidfuzz",
    "unstructured-client",
    "playwright",
    "resend",
    "api-analytics",
    "stripe",
]

[dependency-groups]
dev = [
    "pytest",
    "pytest-asyncio",
    "black",
    "flake8",
    "mypy",
    "pre-commit",
    "jupyter",
    "ipython",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
asyncio_mode = "auto"
filterwarnings = [
    "ignore::DeprecationWarning",
    "ignore::UserWarning",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]
